Move handler to servlets package
